fix(auth): close replay bodies once
