+++ internal/requestconfig/requestconfig.go
@@ -15,7 +15,6 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/openai/openai-go/v3/internal"
@@ -501,22 +500,6 @@ func (b *bodyWithTimeout) Close() error {
 	return err
 }
 
-// closeOnceReadCloser lets Execute clean up bodies when middleware returns an
-// error before reaching the transport, without double-closing bodies that a
-// transport has already closed.
-type closeOnceReadCloser struct {
-	io.ReadCloser
-	once sync.Once
-	err  error
-}
-
-func (b *closeOnceReadCloser) Close() error {
-	b.once.Do(func() {
-		b.err = b.ReadCloser.Close()
-	})
-	return b.err
-}
-
 func retryDelay(res *http.Response, retryCount int) time.Duration {
 	// If the backend tells us to wait a certain amount of time, use that value
 	if retryAfterDelay, ok := parseRetryAfterHeader(res); ok {
@@ -598,7 +581,7 @@ func (cfg *RequestConfig) Execute() (err error) {
 
 		req := cfg.Request.Clone(ctx)
 		if req.Body != nil {
-			req.Body = &closeOnceReadCloser{ReadCloser: req.Body}
+			req.Body = internal.NewCloseOnceReadCloser(req.Body)
 		}
 		if shouldSendRetryCount {
 			req.Header.Set("X-Stainless-Retry-Count", strconv.Itoa(retryCount))
